+++ src/fixtures/brandCarFixtures.js
@@ -1,0 +1,31 @@
+export const BRAND_CAR = {
+    "status": "ok",
+    "data":
+        [
+            {
+                "id": 1,
+                "title": "Audi",
+                "logoFilename": "audi.png"
+            },
+            {
+                "id": 2,
+                "title": "BMW",
+                "logoFilename": "bmw.png"
+            },
+            {
+                "id": 3,
+                "title": "Ford",
+                "logoFilename": "ford.png"
+            },
+            {
+                "id": 4,
+                "title": "Porsche",
+                "logoFilename": "porsche.png"
+            },
+            {
+                "id": 5,
+                "title": "Fiat",
+                "logoFilename": "fiat.png"
+            }
+        ]
+}
